packet: first stab at cluster autoscaler
This adds a `cluster_autoscaler` option to the packet terraform module.

When set to true, it will add and configure the Packet Cluster
Autoscaler in the resulting cluster, which extends or shrinks the worker
node pool depending on the Pod load.

This also adds a number of new variables, some should be taken from the
worker module so they are added as outputs there.

* `worker_user_data`: it's needed for the autoscaler to spawn new nodes,
should be taken from the worker pool module.
* `àuth_token`: the autoscaler needs the auth token to spawn new nodes.
* `worker_type`, `worker_os_channel`: to configure the nodes that will
be spawned by the autoscaler.
* `pool_name`: autoscaler pool name, for now only one pool is supported
and should be taken from the worker module.
* `min_workers`, `mas_workers`: limits for the total number of workers
when using the autoscaler.

Worker nodes created when the cluster is spawned are also part of
the autoscaler pool and are tagged accordingly.

I decided to consider it a core component and add it at cluster set up
but this is up for discussion.

The autoscaler container image is not the `mist` one because of one bug
I found when scaling up the cluster by more than one node. I commented
on the upstrem PR so hopefully this gets fixed in their image soon.

Note that HostEndpoint objects that secure worker nodes are not applied
here, this needs to be solved later.