packet: Merge RAID setup function
Currently there is a generic RAID setup function which creates RAID for
all the disks and another function that creates RAID setup depending on
the disk type. This commit merges the two functions into one.

Also this commit removes the use of sort bash tool and uses `lsblk`'s
flags to sort the results on the basis of size.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>