packet: remove depedency on Route53.
Remove the needed for Route 53 and allow the user to setup the DNS entries by hand.

Terraform will halt the execution and print a message indicating the DNS
entries the user must create before continuing, after the user has set the entries
it can re-apply the plan, Terraform will automatically detect that the entries
are created and continue with the installation.