packet: allow starting kubelet with --cloud-provider=external
To use CCM with Kubernetes we need to pass --cloud-provider=external
to Kubelet. We abstract this with variable "use_ccm" as it makes the
distribution more accessible without having specific knowledge about
CCM requirements in general.

The change is limited in scope to Packet but it is indifferent from
needs for CCM on other cloud providers.

Signed-off-by: Martin Polednik <m.polednik@gmail.com>